Rethink of mutables and contexts. There are now three context types: * ReactiveContext * ReactiveContextMut * SetupContext
Also, mutables no longer have a separate 'setter', but they
can get a getter via `signal()`.

+++ crates/obsidian_ui/src/viewport.rs
@@ -39,7 +39,7 @@ pub fn update_viewport_inset(
             let window = windows.single();
             let ww = window.resolution.physical_width() as f32;
             let wh = window.resolution.physical_height() as f32;
-            let sf = window.resolution.scale_factor();
+            let sf = window.resolution.scale_factor() as f32;
 
             inset.left = min.x;
             inset.top = min.y;
@@ -68,7 +68,7 @@ pub fn update_camera_viewport(
     let window = windows.single();
     let ww = window.resolution.physical_width() as f32;
     let wh = window.resolution.physical_height() as f32;
-    let sf = window.resolution.scale_factor();
+    let sf = window.resolution.scale_factor() as f32;
     let left = viewport_inset.left * sf;
     let right = viewport_inset.right * sf;
     let top = viewport_inset.top * sf;
